+++ spec/stack/stack_updater_spec.rb
@@ -58,11 +58,11 @@ describe SimpleDeploy::StackUpdater do
                                                    :template_body => @template_body,
                                                    :entry         => entry_mock
 
-    stack_updater.update_stack_if_changes( [ { 'param1' => 'new-value' } ] ).
+    stack_updater.update_stack( [ { 'param1' => 'new-value' } ] ).
                   should == true
   end
 
-  it "should update the stack when the template body changes and stack is stable" do
+  it "should update the stack when only the template body changes and stack is stable" do
     attributes = { "param1" => "value1", "param3" => "value3" }
     entry_mock = mock 'entry mock'
     status_mock = mock 'status mock'
@@ -81,10 +81,10 @@ describe SimpleDeploy::StackUpdater do
                       and_return status_mock
     status_mock.should_receive(:wait_for_stable).and_return true
     stack_updater = SimpleDeploy::StackUpdater.new :name          => 'test-stack',
-                                                   :template_body => @template_body,
+                                                   :template_body => @new_template_body,
                                                    :entry         => entry_mock
 
-    stack_updater.update_stack_if_changes([], @new_template_body).should == true
+    stack_updater.update_stack([]).should == true
   end
 
   it "should raise an error when parameters change and stack is not stable" do
@@ -102,20 +102,7 @@ describe SimpleDeploy::StackUpdater do
                                                    :template_body => @template_body,
                                                    :entry         => entry_mock
 
-    lambda {stack_updater.update_stack_if_changes( [ { 'param1' => 'new-value' } ] ) }.
+    lambda {stack_updater.update_stack( [ { 'param1' => 'new-value' } ] ) }.
                   should raise_error
   end
-
-  it "should not update the stack when parameters and template_body don't change" do
-    attributes = { "param3" => "value3" }
-    entry_mock = mock 'entry mock'
-    SimpleDeploy::AWS::CloudFormation.should_receive(:new).exactly(0).times
-    stack_updater = SimpleDeploy::StackUpdater.new :name          => 'test-stack',
-                                                   :template_body => @template_body,
-                                                   :entry         => entry_mock
-
-    stack_updater.update_stack_if_changes( [ { 'another-param' => 'new-value' } ] ).
-                  should == false
-  end
-
 end